Task 2.14 completed

+++ part2/b-phonebook/src/components/Delete.jsx
@@ -1,12 +1,18 @@
 import NoteServices from "../services/NoteServices";
 
-const Delete = ({ id, setPersons }) => {
+const Delete = ({ id, name, setPersons }) => {
   const handleDelete = () => {
+   const confirm = window.confirm(`Delete ${name} ?`);
+   if(confirm){
     NoteServices.deleteNote(id)
       .then(setPersons((prev) => prev.filter((p) => p.id !== id)))
       .catch((error) => {
         console.error("Error deleting note:", error);
       });
+      console.log("Delete success");
+    }else{
+       console.log("Delete cancelled");
+    }
   };
   return (
     <>
